base nueva y termine eliminacion de imagenes

+++ page/src/controllers/product.js
@@ -1,11 +1,5 @@
-
-const product = require("../models/product")
-const color = require('../models/color');
 const { validationResult } = require('express-validator');
-const talle = require('../models/talles');
-const equipos = require('../models/equipos');
-const compras = require('../models/compras');
-const { filterColors, filterByColors } = require("../models/product");
+
 const path = require('path')
 const fs = require('fs')
 const sequelize = require('sequelize')
@@ -58,14 +52,17 @@ module.exports = {
         if (!resultValidation.isEmpty()) {
             try {
                 if(req.files){
-                    let imagen_frente = path.resolve(__dirname,"../../public/uploads",req.file.filename[0])
-                    let imagen_espalda = path.resolve(__dirname,"../../public/uploads",req.file.filename[1])
+                    let imagen_frente = path.resolve(__dirname,"../../public/uploads",req.files[0].filename)
                     if(fs.existsSync(imagen_frente)) {
                         fs.unlinkSync(imagen_frente)
                     }
-                    if(fs.existsSync(imagen_espalda)) {
-                        fs.unlinkSync(imagen_espalda)
+                    if(req.files[1]){
+                        let imagen_espalda = path.resolve(__dirname,"../../public/uploads",req.files[1].filename)
+                         if(fs.existsSync(imagen_espalda)) {
+                            fs.unlinkSync(imagen_espalda)
+                        }
                     }
+                   
                 }
                 return res.render("products/create", {
                     styles:"editar.css",
@@ -78,8 +75,16 @@ module.exports = {
                 res.send(error)
             }
         }
-        let imagenFrente = req.files != undefined ? req.files.find(archivo => archivo.fieldname == 'frente') : imagenDefault;
-        let imagenEspalda = req.files != undefined ? req.files.find(archivo => archivo.fieldname == 'espalda') : imagenDefault;
+        let productoAnterior = await db.Product.findByPk(req.params.id)
+
+        let imagenFrente = path.resolve(__dirname,"../../public/uploads",productoAnterior.imagen_frente)
+        let imagenEspalda = path.resolve(__dirname,"../../public/uploads",productoAnterior.imagen_espalda)
+        if(fs.existsSync(imagenFrente)) {
+            fs.unlinkSync(imagenFrente)
+        }
+        if(fs.existsSync(imagenEspalda)) {
+            fs.unlinkSync(imagenEspalda)
+        }
         let productData =  {
             jugador: req.body.jugador,
             equipo: req.body.equipo,
@@ -104,16 +109,6 @@ module.exports = {
     },
     edit: async (req,res) => {
         try{
-            if(req.files){
-                let imagen_frente = path.resolve(__dirname,"../../public/uploads",req.file.filename[0])
-                let imagen_espalda = path.resolve(__dirname,"../../public/uploads",req.file.filename[1])
-                if(fs.existsSync(imagen_frente)) {
-                    fs.unlinkSync(imagen_frente)
-                }
-                if(fs.existsSync(imagen_espalda)) {
-                    fs.unlinkSync(imagen_espalda)
-                }
-            }
             return res.render("products/edit", {
                 styles:"editar.css",
                 colors: await db.Color.findAll(),
@@ -129,6 +124,19 @@ module.exports = {
         
         if (!resultValidation.isEmpty()) {
             try {
+                if(req.files){
+                    let imagen_frente = path.resolve(__dirname,"../../public/uploads",req.files[0].filename)
+                    if(fs.existsSync(imagen_frente)) {
+                        fs.unlinkSync(imagen_frente)
+                    }
+                    if(req.files[1]){
+                        let imagen_espalda = path.resolve(__dirname,"../../public/uploads",req.files[1].filename)
+                         if(fs.existsSync(imagen_espalda)) {
+                            fs.unlinkSync(imagen_espalda)
+                        }
+                    }
+                   
+                }
                 return res.render("products/edit", {
                     styles:"editar.css",
                     errors: resultValidation.mapped(),
